pyproject.toml
[project]
name = "mae"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy",
    "numpy-stl",
    "scipy",
    "torch>=2.6.0",
    "torchvision",
    "torchaudio",
    "transformers>=4.53.0",
    "healpy>=1.18.1",
    "timm>=1.0.16",
    "tqdm>=4.67.1",
]
